+++ src/libraries/market.js
@@ -42,6 +42,7 @@ import inventoryBadgeManager from '../features/inventory/inventory-badge-manager
 import inventorySort from '../features/inventory/inventory-sort.js';
 import inventoryBadgePrices from '../features/inventory/inventory-badge-prices.js';
 import dungeonTokenTooltips from '../features/inventory/dungeon-token-tooltips.js';
+import treasureTracker from '../features/inventory/treasure-tracker.js';
 import autoAllButton from '../features/inventory/auto-all-button.js';
 import inventoryCategoryTotals from '../features/inventory/inventory-category-totals.js';
 import customTabsFeature from '../features/inventory/custom-tabs/custom-tabs-feature.js';
@@ -84,6 +85,7 @@ toolashaRoot.Market = {
     inventorySort,
     inventoryBadgePrices,
     dungeonTokenTooltips,
+    treasureTracker,
     autoAllButton,
     inventoryCategoryTotals,
     customTabsFeature,